Refactor environment setup functions into config.py
This will allow us to make the environment handling simpler
and more consistent between the local dev server and the
lambda configurations.
