+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": null,
     "lines": null
   },
-  "generated_at": "2020-05-18T11:41:43Z",
+  "generated_at": "2020-05-27T11:57:56Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -46,21 +46,12 @@
     }
   ],
   "results": {
-    "run.py": [
-      {
-        "hashed_secret": "8318df9ecda039deac9868adf1944a29a95c7114",
-        "is_secret": false,
-        "is_verified": false,
-        "line_number": 36,
-        "type": "Secret Keyword"
-      }
-    ],
     "tests/test_main.py": [
       {
         "hashed_secret": "dea742e166979027ae70b28e0a9006fb1010e760",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 237,
+        "line_number": 236,
         "type": "Secret Keyword"
       }
     ]
